% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{censored_keywords}
\alias{censored_keywords}
\title{Censored Keywords.}
\format{A \link[tibble]{tibble} with 26,803 rows and 3 variables:
\describe{
  \item{title}{ Phrase or keyword censored.}
  \item{blocked_last_30_days}{ Amount of content related to \code{title} that is censored.}
  \item{title}{ Last changed in database.}
}}
\source{
\url{https://en.greatfire.org/}
}
\usage{
censored_keywords
}
\description{
A dataset containing censored keywords in Chinese digital and traditional media.
}
\keyword{datasets}
