% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{censored_urls}
\alias{censored_urls}
\title{Censored URLs.}
\format{A \link[tibble]{tibble} with 88,122 rows and 3 variables:
\describe{
  \item{title}{ URL censored.}
  \item{blocked_last_30_days}{ Amount of content related to \code{title} that is censored.}
  \item{title}{ Last changed in database.}
}}
\source{
\url{https://en.greatfire.org/}
}
\usage{
censored_urls
}
\description{
A dataset containing censored URLs in Chinese digital and traditional media.
}
\keyword{datasets}
