% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\name{search}
\alias{search}
\alias{search_url}
\alias{search_keywords}
\title{Search}
\usage{
search_url(q = "github.com")

search_keywords(q = "winnie")
}
\arguments{
\item{q}{Search query}
}
\description{
Convenience function to search through censored keywords and URLs.
}
